# part: Optimize map space usage with one element